+++ lambda/lambda_handlers/__init__.py
@@ -3,7 +3,5 @@
 from __future__ import absolute_import
 
 from . import log_add
-from . import log_get_all_from_today
 
-__all__ = ['log_add',
-           'log_get_all_from_today']
+__all__ = ['log_add']