+++ documentation/release_notes.md
@@ -2,4 +2,5 @@
 
 ## v1.0.0 (Pending)
 ##### Features
+- Added Lambda handlers.
 - Added protocol buffers.